move media files to resources folder, cleanup common components (wip)
