+++ app/app.global.css
@@ -14,8 +14,13 @@ html,
 }
 
 body {
-  font-family: 'Regular', 'Roboto', cursive, sans-serif;
+  font-size: 0.875rem;
+  font-weight: 400;
+  line-height: 1.45;
+  letter-spacing: 0.01;
+  font-family: 'Regular', Arial, Helvetica, sans-serif;
   overflow-y: hidden;
+  color: rgb(0, 0, 0.85);
 }
 
 ::-webkit-scrollbar {
