+++ app/components/common/index.js
@@ -1,21 +1,20 @@
 import AppError from './AppError';
-import AppLoader from './AppLoader';
 import AppLogo from './AppLogo';
 import AppTabs from './AppTabs';
-import ControlTypes from './ControlTypes';
 import Dot from './Dot';
-import Paper from './Paper';
 import SnackbarContent from './SnackbarContent';
 import Transition from './Transition';
-import Widget from './Widget';
 import HelperText from './HelperText';
+
+import Loader from './loader';
+import Paper from './paper';
+import Widget from './widget';
 
 export {
   AppError,
-  AppLoader,
+  Loader,
   AppLogo,
   AppTabs,
-  ControlTypes,
   Dot,
   Paper,
   Transition,
