+++ app/components/common/loader/index.js
@@ -1,0 +1,3 @@
+import Loader from './Loader';
+
+export { Loader };
