+++ app/components/common/loader/styles.js
