+++ app/components/common/paper/Paper.js
@@ -1,11 +1,10 @@
-/* eslint-disable react/require-default-props */
-
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { withStyles } from '@material-ui/core';
 import Paper from '@material-ui/core/Paper';
-import styles from './styles/paper';
+
+import styles from './styles';
 
 const CustomPaper = (props) => {
   const { classes, className, outlined, squared, children, ...rest } = props;
