+++ app/components/common/paper/index.js
@@ -1,0 +1,3 @@
+import Paper from './Paper';
+
+export { Paper };
