+++ app/components/common/paper/styles.js
