--- app/components/common/styles/controlTypes.js
@@ -1,8 +1,0 @@
-const styles = (theme) => ({
-  formControl: {
-    padding: theme.spacing(2),
-    minWidth: 120,
-  },
-});
-
-export default styles;
