+++ app/components/common/widget/Widget.js
@@ -5,7 +5,8 @@ import { withStyles } from '@material-ui/core';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import Divider from '@material-ui/core/Divider';
-import styles from './styles/widget';
+
+import styles from './styles';
 
 const Widget = ({
   classes,
