+++ app/components/common/widget/index.js
@@ -1,0 +1,3 @@
+import Widget from './Widget';
+
+export { Widget };
