+++ package.json
@@ -160,7 +160,7 @@
     "eslint-import-resolver-webpack": "^0.13.0",
     "eslint-plugin-compat": "^3.8.0",
     "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-jsx-a11y": "^6.4.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-react": "^7.21.5",
